The initial preset filter should be for today, not this week
The events button should be added to the list of routes in the side bar nav, so they can go back to events.

In the gallery view, the labels small medium and large should include the number of people. Anywhere you have this number (in the gallery view), and in the size dropdown, the number should be displayed with a people emoji next to it.

The markers on the map should have their radius' increased up to 44px per apple's ui standards

The pill buttons should be able to be toggled off if you click the again. You should also be able to combine different filters as an "or" functionality. If you toggle off a pill filter when no other is selected, it defaults back to "all"